Created new model, converted to one hot encoding, interviewed RNN
The RNN named itself Down. That's it's name now.

Added an LSTM layer, and trained with one hot encoding. It overfitted a
little bit, so it repeats itself a lot.

Introduced probabilistical character selection isntead of using argmax
to introduce more variation when generating text. With it, it doesn't
really repeat itself that much.